+++ lib/squib/api/save.rb
@@ -1,13 +1,12 @@
-require_relative '../template'
 require_relative '../args/card_range'
 require_relative '../args/hand_special'
+require_relative '../args/output_file'
 require_relative '../args/save_batch'
 require_relative '../args/sheet'
-require_relative '../args/template_file'
-require_relative '../args/output_file'
 require_relative '../args/showcase_special'
+require_relative '../args/sprue_file'
 require_relative '../graphics/save_pdf'
-require_relative '../graphics/save_templated_sheet'
+require_relative '../graphics/save_sprue'
 
 module Squib
   class Deck
@@ -23,15 +22,15 @@ module Squib
     def save_pdf(opts = {})
       range = Args::CardRange.new(opts[:range], deck_size: size)
       sheet = Args::Sheet.new(custom_colors, { file: 'output.pdf' }).load!(opts, expand_by: size, layout: layout, dpi: dpi)
-      tmpl_file = Args::TemplateFile.new.load!(opts, expand_by: size)
+      sprue_file = Args::SprueFile.new.load!(opts, expand_by: size)
 
-      if tmpl_file.template_file.nil?
+      if sprue_file.sprue.nil?
         Graphics::SavePDF.new(self).render_pdf(range, sheet)
       else
-        tmpl = Template.load tmpl_file.template_file, dpi
-        Graphics::SaveTemplatedSheetPDF.new(self, tmpl, sheet).render_sheet(
-          range
-        )
+        tmpl = Sprue.load sprue_file.sprue, dpi
+        Graphics::SaveSpruePDF.
+          new(self, tmpl, sheet).
+          render_sheet(range)
       end
     end
 
@@ -52,15 +51,15 @@ module Squib
       range = Args::CardRange.new(opts[:range], deck_size: size)
       batch = Args::SaveBatch.new.load!(opts, expand_by: size, layout: layout, dpi: dpi)
       sheet = Args::Sheet.new(custom_colors, { margin: 0 }, size).load!(opts, expand_by: size, layout: layout, dpi: dpi)
-      tmpl_file = Args::TemplateFile.new.load!(opts, expand_by: size)
+      sprue_file = Args::SprueFile.new.load!(opts, expand_by: size)
 
-      if tmpl_file.template_file.nil?
+      if sprue_file.sprue.nil?
         render_sheet(range, batch, sheet)
       else
-        tmpl = Template.load tmpl_file.template_file, dpi
-        Graphics::SaveTemplatedSheetPNG.new(self, tmpl, batch).render_sheet(
-          range
-        )
+        tmpl = Sprue.load sprue_file.sprue, dpi
+        Graphics::SaveSpruePNG.
+          new(self, tmpl, batch).
+          render_sheet(range)
       end
     end