+++ spec/sprue_spec.rb
@@ -1,22 +1,23 @@
 require 'spec_helper'
+require 'squib/sprues/sprue'
 
-describe Squib::Template do
-  it 'loads a template' do
-    tmpl = Squib::Template.load(template_file('basic.yml'), 100)
+describe Squib::Sprue do
+  it 'loads a sprue' do
+    tmpl = Squib::Sprue.load(sprue_file('basic.yml'), 100)
     expect(tmpl.sheet_width).to eq(850)
     expect(tmpl.sheet_height).to eq(1100)
     expect(tmpl.card_width).to eq(250)
     expect(tmpl.card_height).to eq(350)
     expect(tmpl.dpi).to eq(100)
     expect(tmpl.crop_line_overlay).to eq(
-      Squib::Template::DEFAULTS['crop_line']['overlay']
+      Squib::Sprue::DEFAULTS['crop_line']['overlay']
     )
     expect(tmpl.crop_lines).to eq([])
     expect(tmpl.cards).to eq([{ 'x' => 50, 'y' => 100, 'rotate' => 0 }])
   end
 
   it 'loads from the default templates if none exists' do
-    tmpl = Squib::Template.load('a4_poker_card_9up.yml', 100)
+    tmpl = Squib::Sprue.load('a4_poker_card_9up.yml', 100)
     expect(tmpl.sheet_width).to eq(826.7716527)
     expect(tmpl.sheet_height).to eq(1169.2913373899999)
     expect(tmpl.card_width).to eq(248.03149580999997)
@@ -77,7 +78,7 @@ describe Squib::Template do
   end
 
   it 'loads a template with the coordinates specifying the middle of cards' do
-    tmpl = Squib::Template.load(template_file('card_center_coord.yml'), 100)
+    tmpl = Squib::Sprue.load(sprue_file('card_center_coord.yml'), 100)
     expect(tmpl.sheet_width).to eq(850)
     expect(tmpl.sheet_height).to eq(1100)
     expect(tmpl.card_width).to eq(200)
@@ -94,7 +95,7 @@ describe Squib::Template do
   end
 
   it 'loads a template with customized crop lines' do
-    tmpl = Squib::Template.load(template_file('custom_crop_lines.yml'), 100)
+    tmpl = Squib::Sprue.load(sprue_file('custom_crop_lines.yml'), 100)
     expect(tmpl.sheet_width).to eq(850)
     expect(tmpl.sheet_height).to eq(1100)
     expect(tmpl.card_width).to eq(200)
@@ -130,7 +131,7 @@ describe Squib::Template do
   end
 
   it 'loads a template with rotated cards' do
-    tmpl = Squib::Template.load(template_file('card_rotation.yml'), 100)
+    tmpl = Squib::Sprue.load(sprue_file('card_rotation.yml'), 100)
     expect(tmpl.sheet_width).to eq(850)
     expect(tmpl.sheet_height).to eq(1100)
     expect(tmpl.card_width).to eq(250)
@@ -145,7 +146,7 @@ describe Squib::Template do
 
   it 'fails when sheet_width is not defined' do
     expect do
-      Squib::Template.load(template_file('fail_no_sheet_width.yml'), 100)
+      Squib::Sprue.load(sprue_file('fail_no_sheet_width.yml'), 100)
     end.to raise_error(
       RuntimeError,
       '"sheet_width" is not a String matching /^(\d*[.])?\d+(in|cm|mm)$/'
@@ -154,7 +155,7 @@ describe Squib::Template do
 
   it 'fails when sheet_height is not defined' do
     expect do
-      Squib::Template.load(template_file('fail_no_sheet_height.yml'), 100)
+      Squib::Sprue.load(sprue_file('fail_no_sheet_height.yml'), 100)
     end.to raise_error(
       RuntimeError,
       '"sheet_height" is not a String matching /^(\d*[.])?\d+(in|cm|mm)$/'
@@ -163,7 +164,7 @@ describe Squib::Template do
 
   it 'fails when card_width is not defined' do
     expect do
-      Squib::Template.load(template_file('fail_no_card_width.yml'), 100)
+      Squib::Sprue.load(sprue_file('fail_no_card_width.yml'), 100)
     end.to raise_error(
       RuntimeError,
       '"card_width" is not a String matching /^(\d*[.])?\d+(in|cm|mm)$/'
@@ -172,7 +173,7 @@ describe Squib::Template do
 
   it 'fails when card_height is not defined' do
     expect do
-      Squib::Template.load(template_file('fail_no_card_height.yml'), 100)
+      Squib::Sprue.load(sprue_file('fail_no_card_height.yml'), 100)
     end.to raise_error(
       RuntimeError,
       '"card_height" is not a String matching /^(\d*[.])?\d+(in|cm|mm)$/'