+++ CHANGELOG.md
@@ -4,12 +4,10 @@ Squib follows [semantic versioning](http://semver.org).
 ## v0.14.0 / Unreleased
 
 Features:
+* `save_pdf/save_sheet` method now supports `sprue`, which allows you to define templated layouts and position your cards freely (#217) by @felixleong. See docs for how to use this _very_ powerful feature!
 * `circle` method now supports various `arc` options, so you can draw partial circles (#211) by @sparr
 * `save_sheet` method now supports `rtl` or "right-to-left", for easier duplex printing of backs (#204, #208) by @sparr
-<<<<<<< HEAD
 * `yaml` method for reading in data, much like `csv` and `xlsx` by @blinks
-* `save_pdf/save_sheet` method now supports `template_file`, which allows you to define
-template layouts and position your cards freely (#217) by @felixleong
 
 Special thanks to @sparr, @blinks and @felixleong for all of their work!!